refactor(k8s): update namespace to airflow-elyra-test and add elyra components
This commit changes the namespace from `airflow-elyra` to `airflow-elyra-test` across all Kubernetes resources to support a test environment. Additionally, it introduces Elyra components, including ConfigMaps, PersistentVolumeClaims, and a Deployment, to integrate Elyra with the Airflow setup. The PVC access mode for `airflow-dags-pvc` has also been updated to `ReadWriteOnce` for consistency with other PVCs.